+++ explicaciones.ts
@@ -1,8 +1,8 @@
 import { addEmitHelper } from "typescript";
 import conectarBD from "./db/db";
-import { Enum_Rol, Enum_TipoObjetivo } from "./models/enums";
+import { Enum_Rol, Enum_TipoObjetivo } from "./models/enums/enums";
 import { ObjectiveModel } from "./models/objective";
-import { ProjectModel } from "./models/project";
+import { ProjectModel } from "./models/proyecto/proyecto";
