+++ graphql/resolvers.ts
@@ -1,73 +1,5 @@
-import { ProjectModel } from '../models/project';
-import { UserModel } from '../models/user';
+import { resolversProyecto } from '../models/proyecto/resolvers';
+import { resolversUsuario } from '../models/usuario/resolvers';
+import { resolversAvance } from '../models/avance/resolvers';
 
-const resolvers = {
-  Query: {
-    Usuarios: async (parent, args) => {
-      const usuarios = await UserModel.find();
-      return usuarios;
-    },
-    Usuario: async (parent, args) => {
-      const usuario = await UserModel.findOne({ _id: args._id });
-      return usuario;
-    },
-    Proyectos: async (parent, args) => {
-      const proyectos = await ProjectModel.find().populate('lider');
-      return proyectos;
-    },
-  },
-  
-  Mutation: {
-    crearUsuario: async (parent, args) => {
-      const usuarioCreado = await UserModel.create({
-        nombre: args.nombre,
-        apellido: args.apellido,
-        identificacion: args.identificacion,
-        correo: args.correo,
-        rol: args.rol,
-        estado: args.estado
-      });
-
-      return usuarioCreado;
-    },
-
-    editarUsuario: async (parent, args) => {
-      const usuarioEditado = await UserModel.findByIdAndUpdate(args._id, {
-        nombre: args.nombre,
-        apellido: args.apellido,
-        identificacion: args.identificacion,
-        correo: args.correo,
-        rol: args.rol,
-        estado: args.estado,
-      });
-
-      return usuarioEditado;
-    },
-
-    eliminarUsuario: async (parent, args) => {
-      if (Object.keys(args).includes('_id')) {
-        const usuarioEliminado = await UserModel.findOneAndDelete({ _id: args._id });
-        return usuarioEliminado;
-      } else if (Object.keys(args).includes('correo')) {
-        const usuarioEliminado = await UserModel.findOneAndDelete({ correo: args.correo });
-        return usuarioEliminado;
-      }
-    },
-
-    crearProyecto: async (parent, args) => {
-      const proyectoCreado = await ProjectModel.create({
-        nombre: args.nombre,
-        estado: args.estado,
-        fase: args.fase,
-        fechaInicio: args.fechaInicio,
-        fechaFin: args.fechaFin,
-        presupuesto: args.presupuesto,
-        lider: args.lider,
-        objetivos: args.objetivos,
-      });
-      return proyectoCreado;
-    },
-  },
-};
-
-export { resolvers };
+export const resolvers = [ resolversProyecto, resolversUsuario, resolversAvance ];
