+++ models/avance/tipos.ts
@@ -1,0 +1,22 @@
+import { gql } from "apollo-server-express";
+
+const tiposAvance = gql`
+    type Avance{
+        _id: ID!
+        fecha: Date!
+        descripcion: String!
+        observaciones: [String]
+        proyecto: Proyecto!
+        creadoPor: Usuario!
+    }
+
+    type Query{
+        Avances: [Avance]
+        filtrarAvance(idProyecto: String!): [Avance]
+    }
+    type Mutation{
+        crearAvance(fecha: Date!, descripcion: String!, proyecto: String!, creadoPor: String!): Avance
+    }
+`;
+
+export {tiposAvance};
