+++ models/enums/tipos.ts
@@ -1,0 +1,35 @@
+import { gql } from "apollo-server-express";
+
+const tiposEnums = gql`
+  enum Enum_EstadoUsuario {
+    PENDIENTE
+    AUTORIZADO
+    NO_AUTORIZADO
+  }
+  enum Enum_Rol {
+    ESTUDIANTE
+    LIDER
+    ADMINISTRADOR
+  }
+  enum Enum_EstadoProyecto {
+    ACTIVO
+    INACTIVO
+  }
+  enum Enum_FaseProyecto {
+    INICIADO
+    DESARROLLO
+    TERMINADO
+    NULO
+  }
+  enum Enum_TipoObjetivo {
+    GENERAL
+    ESPECIFICO
+  }
+  enum Enum_EstadoInscripcion {
+    ACEPTADO
+    RECHAZADO
+    PENDIENTE
+  }
+`;
+
+export { tiposEnums };
