+++ models/inscripcion/inscripcion.ts
@@ -1,7 +1,7 @@
 import { Schema, model } from 'mongoose';
-import { Enum_EstadoInscripcion } from './enums';
-import { ProjectModel } from './project';
-import { UserModel } from './user';
+import { Enum_EstadoInscripcion } from '../enums/enums';
+import { ProjectModel } from '../proyecto/proyecto';
+import { UserModel } from '../usuario/usuario';
 
 interface Inscription {
   estado: Enum_EstadoInscripcion;
@@ -37,6 +37,6 @@ const inscriptionSchema = new Schema<Inscription>({
   },
 });
 
-const InscriptionModel = model('Inscripcion', inscriptionSchema);
+const InscriptionModel = model('Inscription', inscriptionSchema);
 
 export { InscriptionModel };
